Begin setting up flask
