+++ pyproject.toml
@@ -21,6 +21,7 @@ python-dotenv = "^1.0.1"
 
 [tool.poetry.group.dev.dependencies]
 pytest-mock = "^3.14.0"
+flask = "^3.0.3"
 
 [build-system]
 requires = ["poetry-core"]
