+++ src/static/style.css
@@ -1,0 +1,16 @@
+body {
+    font-family: Arial, sans-serif;
+}
+
+form {
+    margin: 20px 0;
+}
+
+label {
+    display: inline-block;
+    width: 150px;
+}
+
+input {
+    margin-bottom: 10px;
+}
