+++ src/templates/index.html
@@ -1,0 +1,43 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Transaction Data Visualizer</title>
+</head>
+<body>
+    <h1>Transaction Data Visualizer</h1>
+    <form method="POST" action="/">
+        <label for="account_name">Account Name:</label>
+        <select id="account_name" name="account_name">
+            {% for account_name, account_id in accounts.items() %}
+                <option value="{{ account_name }}">{{ account_name }}</option>
+            {% endfor %}
+        </select>
+
+        <label for="feature_choice">Feature (totals or plot):</label>
+        <select id="feature_choice" name="feature_choice">
+            <option value="totals">Totals</option>
+            <option value="plot">Plot</option>
+        </select>
+
+        <label for="plot_type">Plot Type (bar, line, scatter, pie):</label>
+        <select id="plot_type" name="plot_type">
+            <option value="bar">Bar</option>
+            <option value="line">Line</option>
+            <option value="scatter">Scatter</option>
+            <option value="pie">Pie</option>
+        </select>
+
+        <label for="since">Start Date (YYYY-MM-DD):</label>
+        <input type="date" id="since" name="since">
+
+        <label for="until">End Date (YYYY-MM-DD):</label>
+        <input type="date" id="until" name="until">
+
+        <button type="submit">Submit</button>
+    </form>
+
+    <div id="result"></div>
+</body>
+</html>
